Move dispatcher to top level
